backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
python-dotenv>=1.0.0
google-generativeai>=0.3.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
aiosqlite>=0.19.0
python-multipart>=0.0.6
httpx>=0.25.0
